+++ crates/pm/src/helper/mod.rs
@@ -2,6 +2,7 @@ pub mod auto_update;
 pub mod deps;
 pub mod fs;
 pub mod fuzzy_select;
+pub mod git;
 pub mod global_bin;
 pub mod install_runtime;
 pub mod lock;
